iDealOBot/src/requirements.txt
tabulate 
requests 
beautifulsoup4
selenium
pyodbc
azure-identity
fastapi
uvicorn[standard]
pydantic